+++ Raze.Script.Core/Statements/Expressions/LiteralExpressions/IntegerLiteralExpression.cs
@@ -4,9 +4,9 @@ internal class IntegerLiteralExpression : LiteralExpression
 {
     public override object Value => _value;
 
-    private readonly int _value;
+    private readonly Int128 _value;
 
-    public IntegerLiteralExpression(int value, int startLine, int startColumn)
+    public IntegerLiteralExpression(Int128 value, int startLine, int startColumn)
         : base(startLine, startColumn)
     {
         _value = value;
